+++ app/src/main/java/com/oli/testapp/MainActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.oli.testapp.feature.presentation.CoinListScreen
+import com.oli.testapp.feature.presentation.components.CoinListScreen
 import com.oli.testapp.feature.presentation.CoinListViewModel
 import com.oli.testapp.ui.theme.TestAppTheme
 import dagger.hilt.android.AndroidEntryPoint
