+++ app/src/main/java/com/oli/testapp/feature/presentation/components/CoinListItem.kt
@@ -1,4 +1,4 @@
-package com.oli.testapp.feature.presentation
+package com.oli.testapp.feature.presentation.components
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
