+++ app/src/main/java/com/oli/testapp/feature/presentation/components/CoinListScreen.kt
@@ -1,4 +1,4 @@
-package com.oli.testapp.feature.presentation
+package com.oli.testapp.feature.presentation.components
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -9,6 +9,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import com.oli.testapp.feature.presentation.CoinListViewModel
 import kotlinx.coroutines.flow.collectLatest
 
 @Composable
